requirements.txt
fastapi
motor
pydantic
python-dotenv
pytest
httpx